refactor(update-status): extact login into hooks
